requirements.txt
duckdb
pandas
pyyaml
python-dotenv
openpyxl
matplotlib
streamlit
